agentic_ai/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12
pypdf==4.3.1
reportlab==4.2.5
pandas==2.2.3
matplotlib==3.9.2
openpyxl==3.1.5
python-pptx==1.0.2
beautifulsoup4==4.12.3
langchain==0.3.4
langchain-core==0.3.12
langgraph==0.2.34
